+++ src/chapters/workshop3/index.js
@@ -1,10 +1,12 @@
 import { introToWorkshop3Chapters } from "./workshop3-intro"
 import { workshop3ConclusionChapters } from "./conclusion"
 import { introToRDSChapters } from "./rds"
+import { rdsCicdChapters} from "./cicd-rds"
 
 
 export const workshop3Chapters = [
   ...introToWorkshop3Chapters,
   ...workshop3ConclusionChapters,
-  ...introToRDSChapters
+  ...introToRDSChapters,
+  ...rdsCicdChapters
 ]
